requirements.txt
aiogram==2.24
geopy==2.3.0
python-dotenv==0.21.1
python-socks==2.1.1
requests==2.28.2
